chore(team): split remote relay delivery
